refactor(natural-language-understanding): remove compatibility layer for natural language understanding
BREAKING CHANGE: The `version_date` parameter in Natural Language Understanding is no longer supported.

To migrate your code, use the parameter name `version` instead.
